+++ src/ksi/log.c
@@ -88,7 +88,7 @@ KSI_LOG_FN(info, INFO);
 KSI_LOG_FN(error, ERROR);
 KSI_LOG_FN(fatal, FATAL);
 
-int KSI_LOG_logBlob(KSI_CTX *ctx, int level, const char *prefix, const unsigned char *data, unsigned data_len) {
+int KSI_LOG_logBlob(KSI_CTX *ctx, int level, const char *prefix, const unsigned char *data, size_t data_len) {
 	int res = KSI_UNKNOWN_ERROR;
 	char *logStr = NULL;
 	size_t logStr_size = 0;
